pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = ""
authors = [
    {name = "Brian McKean",email = "brian@rethlasoft.com"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "openai-whisper (>=20250625,<20250626)",
    "whisperx (>=3.7.4,<4.0.0)",
    "anthropic (>=0.74.1,<0.75.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "simple-diarizer (>=0.0.13,<0.0.14)",
    "langfuse (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
